fgkaslr: reconstruct kallsyms
kallsyms is generated at build time and must contain a sorted
array of offsets as well as a table of indices into a compressed
token table. We don't need to regenerate the tokens, but we will
need to resort both the offsets and the names tables after we've
randomized everything. In order to do this, the boot kernel needs
to get access to the address of these tables so it can rewrite
them in place.  This patch adds an option to retain a few symbols
during the objcopy so that we can read their location in the boot
kernel. We use the existing kernel sort algorithm to resort
the offsets and names after randomization but before relocation.

Signed-off-by: Kristen Carlson Accardi <kristen@linux.intel.com>