percpu: fix first chunk size calculation for populated bitmap
Variable populated, which is a member of struct pcpu_chunk, is used as a
unit of size of unsigned long.
However, size of populated is miscounted. So, I fix this minor part.

The first chunk is special as it holds a preconfigured number of dynamic
pages (generally 7). However, the first chunk size calculation only
added a single byte (instead of # of longs) to hold the populated
bitmap. The dynamic page configuration is not a very often used tunable,
so I attribute this to why we aren't seeing a more widespread issue.

Signed-off-by: Sunghyun Jin <mcsmonk@gmail.com>
Signed-off-by: Dennis Zhou <dennis@kernel.org>
[Added explanation why this isn't a bigger issue.]